chore: cleanup (#149)
