+++ .github/workflows/poetry-test.yml
@@ -22,7 +22,7 @@ jobs:
         uses: snok/install-poetry@v1
 
       - name: Install python dependencies
-        run: poetry install
+        run: poetry install --with gdino
 
       - name: Install ROS 2 dependencies
         shell: bash
