+++ pyproject.toml
@@ -46,6 +46,7 @@ pypdf = "^4.2.0"
 langchain-ollama = "^0.1.1"
 
 streamlit = "^1.37.1"
+deprecated = "^1.2.14"
 [tool.poetry.group.dev.dependencies]
 ipykernel = "^6.29.4"
