+++ src/rai/package.xml
@@ -13,6 +13,10 @@
   <test_depend>ament_pep257</test_depend>
   <test_depend>python3-pytest</test_depend>
 
+  <exec_depend>nav2_msgs</exec_depend>
+  <exec_depend>nav2_simple_commander</exec_depend>
+  <exec_depend>tf_transformations</exec_depend>
+
   <export>
     <build_type>ament_python</build_type>
   </export>
