+++ src/rai/rai/cli/rai_cli.py
@@ -23,7 +23,8 @@ from langchain_core.messages import SystemMessage
 from langchain_openai import ChatOpenAI, OpenAIEmbeddings
 
 from rai.apps.talk_to_docs import ingest_documentation
-from rai.scenario_engine.messages import HumanMultimodalMessage, preprocess_image
+from rai.messages import preprocess_image
+from rai.messages.multimodal import HumanMultimodalMessage
 
 
 def parse_whoami_package():
