--- src/rai/rai/communication/README.md
@@ -1,35 +1,0 @@
-# 📘 README for Communication Modules
-
-## 📨 Communication.py
-
-### Overview
-
-This module is designed to handle general communication tasks like sending emails. It's set up to use SMTP (Simple Mail Transfer Protocol) for email operations, allowing the application to send notifications, alerts, or any communication via email.
-
-### What to Implement
-
-- **Add More Communication Methods**: Beyond email, you might want to integrate other forms of communication like SMS, direct messaging services, or even automated phone calls.
-- **Security Enhancements**: Implement more robust security measures for handling credentials and securing communication channels.
-- **Error Handling**: Enhance error handling to manage network issues or authentication errors more gracefully.
-
-### Currently Implemented
-
-- **Email Sending**: Setup to send emails using SMTP with attachments. It includes basic error handling for SMTP authentication errors and checks for missing email credentials.
-
-## 🤖 ros_communication.py
-
-### Overview
-
-This file is specifically tailored for communication within ROS (Robot Operating System) environments. It deals with subscribing to ROS topics, waiting for messages, and processing those messages. It's crucial for robotic applications where real-time data handling and sensor integration are required.
-
-### What to Implement
-
-- **Expand Topic Handling**: Include subscriptions to more diverse topic types and handle different data formats coming from various sensors or inputs.
-- **Integration with More ROS Versions**: Ensure compatibility with different versions of ROS or other robotic middleware.
-- **Enhanced Data Processing**: Implement more complex data processing functions that can convert incoming data into more usable formats or derive more insights.
-
-### Currently Implemented
-
-- **Message Subscription and Retrieval**: Functions to wait for and retrieve messages from specified ROS topics.
-- **Image Data Handling**: Includes a specialized class for grabbing image data from a ROS topic, converting it from ROS image formats to standard encodings.
-- **Utilities for ROS Entities**: Functions to list available topics, nodes, and services in the ROS environment, used for usage agnostic dynamic LLM systems.
