+++ src/rai/rai/tools/__init__.py
@@ -12,14 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-from .hmi_tools import PlayVoiceMessageTool, SendEmailTool, WaitForSecondsTool
-from .planning_tools import AddTaskTool, GetNewTaskTool
-
-__all__ = [
-    "PlayVoiceMessageTool",
-    "SendEmailTool",
-    "WaitForSecondsTool",
-    "AddTaskTool",
-    "GetNewTaskTool",
-]
