+++ src/rai/rai/tools/ros/__init__.py
@@ -13,18 +13,7 @@
 # limitations under the License.
 #
 
-from .cat_demo_tools import (
-    ContinueActionTool,
-    ReplanWithoutCurrentPathTool,
-    UseHonkTool,
-    UseLightsTool,
-)
 from .cli import Ros2InterfaceTool, Ros2ServiceTool, Ros2TopicTool
-from .mock_tools import (
-    ObserveSurroundingsTool,
-    OpenSetSegmentationTool,
-    VisualQuestionAnsweringTool,
-)
 from .tools import (
     AddDescribedWaypointToDatabaseTool,
     GetCurrentPositionTool,
@@ -32,13 +21,6 @@ from .tools import (
 )
 
 __all__ = [
-    "ReplanWithoutCurrentPathTool",
-    "UseHonkTool",
-    "UseLightsTool",
-    "ContinueActionTool",
-    "OpenSetSegmentationTool",
-    "VisualQuestionAnsweringTool",
-    "ObserveSurroundingsTool",
     "Ros2TopicTool",
     "Ros2InterfaceTool",
     "Ros2ServiceTool",
